added sass and divided app into components and more reasonable structure
