+++ package.json
@@ -19,6 +19,7 @@
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@vitejs/plugin-react": "^4.3.1",
+    "sass": "^1.77.5",
     "typescript": "^5.4.5",
     "vite": "^5.3.1"
   }
