+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App.tsx';
-import './index.css';
+import App from './components/App.tsx';
+import './styles/style.scss'
 
 
 const container = document.getElementById('root');
