+++ src/variables.ts
@@ -1,0 +1,226 @@
+const genres = [
+  {value: ""},
+  {value: "Action"},
+  {value: "Adult"},
+  {value: "Adventure"},
+  {value: "Anime"},
+  {value: "Children"},
+  {value: "Comedy"},
+  {value: "Crime"},
+  {value: "DIY"},
+  {value: "Drama"},
+  {value: "Espionage"},
+  {value: "Family"},
+  {value: "Fantasy"},
+  {value: "Food"},
+  {value: "History"},
+  {value: "Horror"},
+  {value: "Legal"},
+  {value: "Medical"},
+  {value: "Music"},
+  {value: "Mystery"},
+  {value: "Nature"},
+  {value: "Romance"},
+  {value: "Science-Fiction"},
+  {value: "Sports"},
+  {value: "Supernatural"},
+  {value: "Thriller"},
+  {value: "Travel"},
+  {value: "War"},
+  {value: "Western"}
+];
+
+const types = [
+  {value: ""},
+  {value: "Scripted"},
+  {value: "Animation"},
+  {value: "Reality"},
+  {value: "Talk Show"},
+  {value: "Documentary"},
+  {value: "Game Show"},
+  {value: "News"},
+  {value: "Sports"},
+  {value: "Variety"},
+  {value: "Award Show"},
+  {value: "Panel Show"}
+];
+
+const statuses = [
+  {value: ""},
+  {value: "Running"},
+  {value: "Ended"},
+  {value: "To Be Determined"},
+  {value: "In Development"},
+];
+
+const languages = [
+  {value: ""},
+  {value: "Afrikaans"},
+  {value: "Albanian"},
+  {value: "Arabic"},
+  {value: "Armenian"},
+  {value: "Azerbaijani"},
+  {value: "Basque"},
+  {value: "Belarusian"},
+  {value: "Bengali"},
+  {value: "Bosnian"},
+  {value: "Bulgarian"},
+  {value: "Catalan"},
+  {value: "Chechen"},
+  {value: "Chinese"},
+  {value: "Croatian"},
+  {value: "Czech"},
+  {value: "Danish"},
+  {value: "Divehi"},
+  {value: "Dutch"},
+  {value: "English"},
+  {value: "Estonian"},
+  {value: "Finnish"},
+  {value: "French"},
+  {value: "Galician"},
+  {value: "Georgian"},
+  {value: "German"},
+  {value: "Greek"},
+  {value: "Gujarati"},
+  {value: "Hebrew"},
+  {value: "Hindi"},
+  {value: "Hungarian"},
+  {value: "Iceland"},
+  {value: "Indonesian"},
+  {value: "Irish"},
+  {value: "Italian"},
+  {value: "Japanese"},
+  {value: "Javanese"},
+  {value: "Kannada"},
+  {value: "Kazakh"},
+  {value: "Kongo"},
+  {value: "Korean"},
+  {value: "Latin"},
+  {value: "Latvian"},
+  {value: "Lithuanian"},
+  {value: "Luxembourgish"},
+  {value: "Malay"},
+  {value: "Malayalam"},
+  {value: "Marathi"},
+  {value: "Mongolian"},
+  {value: "Norwegian"},
+  {value: "Panjabi"},
+  {value: "Pashto"},
+  {value: "Persian"},
+  {value: "Polish"},
+  {value: "Portuguese"},
+  {value: "Romanian"},
+  {value: "Russian"},
+  {value: "Serbian"},
+  {value: "Sinhalese"},
+  {value: "Slovak"},
+  {value: "Slovenian"},
+  {value: "Spanish"},
+  {value: "Swedish"},
+  {value: "Tagalog"},
+  {value: "Tamil"},
+  {value: "Telugu"},
+  {value: "Thai"},
+  {value: "Turkish"},
+  {value: "Ukrainian"},
+  {value: "Urdu"},
+  {value: "Uzbek"},
+  {value: "Vietnamese"},
+  {value: "Welsh"},
+  {value: "Scottish Gaelic"}
+];
+
+const countries = [
+  {value: ""},
+  {value: "Afghanistan"},
+  {value: "Albania"},
+  {value: "Algeria "},
+  {value: "Argentina"},
+  {value: "Armenia"},
+  {value: "Australia"},
+  {value: "Austria"},
+  {value: "Azerbaijan"},
+  {value: "Bangladesh"},
+  {value: "Belarus"},
+  {value: "Belgium"},
+  {value: "Bosnia and Herzegovina"},
+  {value: "Brazil"},
+  {value: "Bulgaria"},
+  {value: "Canada"},
+  {value: "Chile"},
+  {value: "China"},
+  {value: "Colombia"},
+  {value: "Croatia"},
+  {value: "Cyprus"},
+  {value: "Czech Republic"},
+  {value: "Denmark"},
+  {value: "Egypt"},
+  {value: "Estonia"},
+  {value: "Faroe Islands"},
+  {value: "Finland"},
+  {value: "France"},
+  {value: "French Polynesia"},
+  {value: "Georgia"},
+  {value: "Germany"},
+  {value: "Greece"},
+  {value: "Hong Kong"},
+  {value: "Hungary"},
+  {value: "Iceland"},
+  {value: "India"},
+  {value: "Indonesia"},
+  {value: "Iran, Islamic Republic of"},
+  {value: "Iraq"},
+  {value: "Ireland"},
+  {value: "Israel"},
+  {value: "Italy"},
+  {value: "Japan"},
+  {value: "Kazakhstan"},
+  {value: "Korea, Democratic People's Republic of"},
+  {value: "Korea, Republic of"},
+  {value: "Latvia"},
+  {value: "Lebanon"},
+  {value: "Lithuania"},
+  {value: "Luxembourg"},
+  {value: "Malaysia"},
+  {value: "Maldives"},
+  {value: "Mexico"},
+  {value: "Moldova, Republic of"},
+  {value: "Mongolia"},
+  {value: "Netherlands"},
+  {value: "New Zealand"},
+  {value: "Nigeria"},
+  {value: "Norway"},
+  {value: "Pakistan"},
+  {value: "Peru"},
+  {value: "Philippines"},
+  {value: "Poland"},
+  {value: "Portugal"},
+  {value: "Puerto Rico"},
+  {value: "Qatar"},
+  {value: "Romania"},
+  {value: "Russian Federation"},
+  {value: "Saudi Arabia"},
+  {value: "Serbia"},
+  {value: "Singapore"},
+  {value: "Slovakia"},
+  {value: "Slovenia"},
+  {value: "South Africa"},
+  {value: "Spain"},
+  {value: "Sri Lanka"},
+  {value: "Sweden"},
+  {value: "Switzerland"},
+  {value: "Taiwan, Province of China"},
+  {value: "Thailand"},
+  {value: "Trinidad and Tobago"},
+  {value: "Tunisia"},
+  {value: "Turkey"},
+  {value: "Ukraine"},
+  {value: "United Arab Emirates"},
+  {value: "United Kingdom"},
+  {value: "United States"},
+  {value: "Uzbekistan"},
+  {value: "Venezuela, Bolivarian Republic of"},
+  {value: "Viet Nam"}
+];
+
+export { genres, types, statuses, languages, countries };
