+++ tsconfig.json
@@ -23,7 +23,7 @@
     "noUnusedParameters": true
   },
   "include": [
-    "src"
+    "src/**/*.ts", "src/**/*.tsx"
   ],
   "references": [{ "path": "./tsconfig.node.json"}]
 }
