+++ vite.config.ts
@@ -2,5 +2,5 @@ import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
 export default defineConfig({
-    plugins: [react()]
+  plugins: [react()],
 });
